rename new string module
